requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
geopy==2.4.1
requests==2.31.0
rasterio==1.3.9
aiohttp==3.9.1
aiofiles==23.2.1
numpy<2.0
Pillow==10.1.0

